Fixed bug, added small gradient abstraction
The bug was pretty easily fixed by refreshing the camera script at the beginning of a game.

Going through my Start() method for the player, I saw a very easy abstraction for how I was instantiating all of my gradients, so I pulled that out into convenience method and used it to cut down the length of the Start() method.